+++ src/react/components/stepper/customElements/StepperTab.js
@@ -2,18 +2,22 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import XUIProgressCircular from '../../progressindicator/XUIProgressCircular';
 import { NAME_SPACE } from '../helpers/constants';
-import { enrichTabProps } from '../helpers/enrichprops';
 import StepperIcon from './StepperIcon';
 import { ns } from '../../helpers/xuiClassNamespace';
 
 export default class StepperTab extends PureComponent {
+	handleClick = () => {
+		const { handleClick, step } = this.props;
+
+		handleClick(step - 1);
+	};
+
 	render = () => {
 		const {
 			id,
 			name,
 			description,
 			step,
-			handleClick,
 			isError,
 			isComplete,
 			isProgress,
@@ -21,13 +25,13 @@ export default class StepperTab extends PureComponent {
 			currentProgress,
 			tabIndex,
 			linkClasses,
-		} = enrichTabProps(this.props);
+		} = this.props;
 
 		return (
 			<button
 				type="button"
 				className={linkClasses}
-				onClick={handleClick}
+				onClick={this.handleClick}
 				tabIndex={tabIndex}
 			>
 
@@ -75,12 +79,13 @@ StepperTab.propTypes = {
 	id: PropTypes.string.isRequired,
 	name: PropTypes.string.isRequired,
 	description: PropTypes.string,
+	linkClasses: PropTypes.string,
 	step: PropTypes.number,
+	tabIndex: PropTypes.number,
 	handleClick: PropTypes.func,
 	isError: PropTypes.bool,
 	isComplete: PropTypes.bool,
 	isActive: PropTypes.bool,
-	isDisabled: PropTypes.bool,
 	isProgress: PropTypes.bool,
 	totalProgress: PropTypes.number,
 	currentProgress: PropTypes.number,